abd-last-commit

+++ src/main/java/main/funcDeleteObservation.java
@@ -4,15 +4,19 @@ import java.sql.*;
 
 public class funcDeleteObservation {
 
-    public static void deleteObservations() throws SQLException {
-
+    public static int  flagDeleteObservations=0;
+    public static void deleteObservations(int testp) throws SQLException {
+       flagDeleteObservations=0;
         Connection connection = null;
         PreparedStatement pst= null;
         ResultSet rs = null;
 
         connection = DriverManager.getConnection("jdbc:mysql://localhost:3306/Sakan", "root", "");
-        pst = connection.prepareStatement("DELETE FROM system_observation " );
-        pst.executeUpdate();
+        if(testp==1) {
+            pst = connection.prepareStatement("DELETE FROM system_observation ");
+            pst.executeUpdate();
+        }
+    flagDeleteObservations=1;
 
         System.out.println("You have successfully Deleted the Observations");
     }
